fix: not all patches of children are called
It does not work correctly if child nodes are removed,
since live NodeList ($parent.childNodes) updates the collection automatically
when the DOM changes (e.g. child is removed). Hence, forEach is not called for
all (original) elements (child nodes).
For example, if there are 3 child nodes and the first patch keeps the child and
the second and third patch delete the child, then only the first and second
patches are called. The third patch is not called, since the second patch removes
the node, which removes it from $parent.childNodes.